add configurable default cache behavior lambda@edge
