+++ __tests__/custom-url-origin.test.js
@@ -21,8 +21,14 @@ describe('Input origin as a custom url', () => {
     component = await createComponent()
   })
 
-  it('creates distribution with custom url origin', async () => {
+  it('creates distribution with custom url origin and sets defaults', async () => {
     await component.default({
+      defaults: {
+        ttl: 10,
+        'lambda@edge': {
+          'origin-request': 'arn:aws:lambda:us-east-1:123:function:originRequestFunction'
+        }
+      },
       origins: ['https://mycustomorigin.com']
     })
