+++ lib/getDefaultCacheBehavior.js
@@ -1,5 +1,7 @@
+const addLambdaAtEdgeToCacheBehavior = require('./addLambdaAtEdgeToCacheBehavior')
+
 module.exports = (originId, defaults = {}) => {
-  return {
+  const defaultCacheBehavior = {
     TargetOriginId: originId,
     ForwardedValues: {
       QueryString: false,
@@ -40,4 +42,8 @@ module.exports = (originId, defaults = {}) => {
     },
     FieldLevelEncryptionId: ''
   }
+
+  addLambdaAtEdgeToCacheBehavior(defaultCacheBehavior, defaults['lambda@edge'])
+
+  return defaultCacheBehavior
 }
