+++ lib/parseInputOrigins.js
@@ -1,12 +1,7 @@
 const getOriginConfig = require('./getOriginConfig')
 const getCacheBehavior = require('./getCacheBehavior')
+const addLambdaAtEdgeToCacheBehavior = require('./addLambdaAtEdgeToCacheBehavior')
 
-const validLambdaTriggers = [
-  'viewer-request',
-  'origin-request',
-  'origin-response',
-  'viewer-response'
-]
 module.exports = (origins) => {
   const distributionOrigins = {
     Quantity: 0,
@@ -26,23 +21,7 @@ module.exports = (origins) => {
         const pathPatternConfig = origin.pathPatterns[pathPattern]
         const cacheBehavior = getCacheBehavior(pathPattern, pathPatternConfig, originConfig.Id)
 
-        const lambdaAtEdge = pathPatternConfig['lambda@edge'] || {}
-
-        Object.keys(lambdaAtEdge).forEach((eventType) => {
-          if (!validLambdaTriggers.includes(eventType)) {
-            throw new Error(
-              `"${eventType}" is not a valid lambda trigger. See https://docs.aws.amazon.com/AmazonCloudFront/latest/DeveloperGuide/lambda-cloudfront-trigger-events.html for valid event types.`
-            )
-          }
-
-          cacheBehavior.LambdaFunctionAssociations.Quantity =
-            cacheBehavior.LambdaFunctionAssociations.Quantity + 1
-          cacheBehavior.LambdaFunctionAssociations.Items.push({
-            EventType: eventType,
-            LambdaFunctionARN: lambdaAtEdge[eventType],
-            IncludeBody: true
-          })
-        })
+        addLambdaAtEdgeToCacheBehavior(cacheBehavior, pathPatternConfig['lambda@edge'])
 
         distributionCacheBehaviors = {
           Quantity: 0,
